# test/e2e/__snapshots__/test_openapi/test_OpenAPIの形が変わっていないことを確認.json
{
  "components": {
    "schemas": {
...
      "AivmInfo": {
        "description": "AIVM (Aivis Voice Model) 仕様に準拠した音声合成モデルのメタデータ情報\nAIVM マニフェストには、音声合成モデルに関連する全てのメタデータが含まれる\nspeakers フィールド内の話者情報は、VOICEVOX ENGINE との API 互換性のために\nAIVM マニフェストを元に Speaker / SpeakerStyle / SpeakerInfo / StyleInfo モデルに変換したもの",
        "properties": {
          "file_path": {
            "format": "path",
            "title": "AIVMX ファイルのインストール先パス",
            "type": "string"
          },
          "file_size": {
            "title": "AIVMX ファイルのインストールサイズ (バイト単位)",
            "type": "integer"
          },
          "is_loaded": {
            "title": "この音声合成モデルがロードされているかどうか",
            "type": "boolean"
          },
          "is_update_available": {
            "title": "この音声合成モデルの新しいバージョンが AivisHub で公開されているかどうか",
            "type": "boolean"
          },
          "latest_version": {
            "title": "この音声合成モデルの AivisHub で公開されている最新バージョン (AivisHub で公開されていない場合は AIVM マニフェスト記載のバージョン)",
            "type": "string"
          },
          "manifest": {
            "$ref": "#/components/schemas/AivmManifest",
            "title": "AIVM マニフェスト"
          },
          "speakers": {
            "items": {
              "$ref": "#/components/schemas/LibrarySpeaker"
            },
            "title": "話者情報のリスト (VOICEVOX ENGINE 互換)",
            "type": "array"
          }
        },
        "required": [
          "is_loaded",
          "is_update_available",
          "latest_version",
          "file_path",
          "file_size",
          "manifest",
          "speakers"
        ],
        "title": "AivmInfo",
        "type": "object"
      },
      "AivmManifest": {
        "description": "AIVM マニフェストのスキーマ ",
        "properties": {
          "creators": {
            "default": [],
            "items": {
              "maxLength": 255,
              "minLength": 1,
              "type": "string"
            },
            "title": "Creators",
            "type": "array"
          },
          "description": {
            "default": "",
            "maxLength": 140,
            "title": "Description",
            "type": "string"
          },
          "license": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "manifest_version": {
            "const": "1.0",
            "title": "Manifest Version",
            "type": "string"
          },
          "model_architecture": {
            "$ref": "#/components/schemas/ModelArchitecture"
          },
          "model_format": {
            "$ref": "#/components/schemas/ModelFormat"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "speakers": {
            "items": {
              "$ref": "#/components/schemas/AivmManifestSpeaker"
            },
            "title": "Speakers",
            "type": "array"
          },
          "training_epochs": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Epochs"
          },
          "training_steps": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Steps"
          },
          "uuid": {
            "format": "uuid",
            "title": "Uuid",
            "type": "string"
          },
          "version": {
            "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-((?:0|[1-9]\\d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\\.(?:0|[1-9]\\d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\\+([0-9a-zA-Z-]+(?:\\.[0-9a-zA-Z-]+)*))?$",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "manifest_version",
          "name",
          "model_architecture",
          "model_format",
          "uuid",
          "version",
          "speakers"
        ],
        "title": "AivmManifest",
        "type": "object"
      },
      "AivmManifestSpeaker": {
        "description": "AIVM マニフェストの話者情報 ",
        "properties": {
          "icon": {
            "pattern": "^data:image/(jpeg|png);base64,[A-Za-z0-9+/=]+$",
            "title": "Icon",
            "type": "string"
          },
          "local_id": {
            "minimum": 0.0,
            "title": "Local Id",
            "type": "integer"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "styles": {
            "items": {
              "$ref": "#/components/schemas/AivmManifestSpeakerStyle"
            },
            "title": "Styles",
            "type": "array"
          },
          "supported_languages": {
            "items": {
              "pattern": "^[a-z]{2,3}(?:-[A-Z]{4})?(?:-(?:[A-Z]{2}|\\d{3}))?(?:-(?:[A-Za-z0-9]{5,8}|\\d[A-Za-z0-9]{3}))*(?:-[A-Za-z](?:-[A-Za-z0-9]{2,8})+)*(?:-x(?:-[A-Za-z0-9]{1,8})+)?$",
              "type": "string"
            },
            "title": "Supported Languages",
            "type": "array"
          },
          "uuid": {
            "format": "uuid",
            "title": "Uuid",
            "type": "string"
          }
        },
        "required": [
          "name",
          "icon",
          "supported_languages",
          "uuid",
          "local_id",
          "styles"
        ],
        "title": "AivmManifestSpeaker",
        "type": "object"
      },
      "AivmManifestSpeakerStyle": {
        "description": "AIVM マニフェストの話者スタイル情報 ",
        "properties": {
          "icon": {
            "anyOf": [
              {
                "pattern": "^data:image/(jpeg|png);base64,[A-Za-z0-9+/=]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "local_id": {
            "maximum": 31.0,
            "minimum": 0.0,
            "title": "Local Id",
            "type": "integer"
          },
          "name": {
            "maxLength": 20,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "voice_samples": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AivmManifestVoiceSample"
            },
            "title": "Voice Samples",
            "type": "array"
          }
        },
        "required": [
          "name",
          "local_id"
        ],
        "title": "AivmManifestSpeakerStyle",
        "type": "object"
      },
      "AivmManifestVoiceSample": {
        "description": "AIVM マニフェストのボイスサンプル情報 ",
        "properties": {
          "audio": {
            "pattern": "^data:audio/(wav|mp4);base64,[A-Za-z0-9+/=]+$",
            "title": "Audio",
            "type": "string"
          },
          "transcript": {
            "minLength": 1,
            "title": "Transcript",
            "type": "string"
          }
        },
        "required": [
          "audio",
          "transcript"
        ],
        "title": "AivmManifestVoiceSample",
        "type": "object"
      },
...
      "Body_install_model_aivm_models_install_post": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "format": "binary",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "AIVMX ファイル (`.aivmx`)",
            "title": "File"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "AIVMX ファイルの URL",
            "title": "Url"
          }
        },
        "title": "Body_install_model_aivm_models_install_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LibrarySpeaker": {
        "description": "音声ライブラリに含まれるキャラクターの情報",
        "properties": {
          "speaker": {
            "$ref": "#/components/schemas/Speaker"
          },
          "speaker_info": {
            "$ref": "#/components/schemas/SpeakerInfo"
          }
        },
        "required": [
          "speaker",
          "speaker_info"
        ],
        "title": "LibrarySpeaker",
        "type": "object"
      },
...
      "ModelArchitecture": {
        "enum": [
          "Style-Bert-VITS2",
          "Style-Bert-VITS2 (JP-Extra)"
        ],
        "title": "ModelArchitecture",
        "type": "string"
      },
      "ModelFormat": {
        "enum": [
          "Safetensors",
          "ONNX"
        ],
        "title": "ModelFormat",
        "type": "string"
      },
...
      "Preset": {
        "description": "プリセット情報",
        "properties": {
          "id": {
            "title": "プリセット ID",
            "type": "integer"
          },
          "intonationScale": {
            "description": "話者スタイルの声色の強弱を 0.0 ~ 2.0 の範囲で指定する (デフォルト: 1.0) 。\n値が大きいほどそのスタイルに近い抑揚がついた声になる。\n例えば話者スタイルが「うれしい」なら、値が大きいほどより嬉しそうな明るい話し方になる。\n一方、話者やスタイルによっては、数値を上げすぎると発声がおかしくなったり、棒読みで不自然な声になる場合もある。\nちゃんと発声できる「スタイルの強さ」の上限は話者やスタイルによって異なるため、適宜調整が必要。\n全スタイルの平均であるノーマルスタイルには指定できない (値にかかわらず無視される) 。",
            "title": "全体のスタイルの強さ (「全体の抑揚」ではない点で VOICEVOX ENGINE と異なる)",
            "type": "number"
          },
          "name": {
            "title": "プリセット名",
            "type": "string"
          },
          "pauseLength": {
            "description": "句読点などの無音時間。null のときは無視される。デフォルト値は null 。",
            "title": "AivisSpeech Engine ではサポートされていないフィールドです (常に無視されます)",
            "type": "number"
          },
          "pauseLengthScale": {
            "default": 1,
            "description": "句読点などの無音時間（倍率）。デフォルト値は 1 。",
            "title": "AivisSpeech Engine ではサポートされていないフィールドです (常に無視されます)",
            "type": "number"
          },
          "pitchScale": {
            "description": "全体の音高を -0.15 ~ 0.15 の範囲で指定する (デフォルト: 0.0) 。\n値が大きいほど高い声になる。\nVOICEVOX ENGINE と異なり、この値を 0.0 から変更すると音質が劣化するため注意が必要。",
            "title": "全体の音高",
            "type": "number"
          },
          "postPhonemeLength": {
            "title": "音声の後の無音時間 (秒)",
            "type": "number"
          },
          "prePhonemeLength": {
            "title": "音声の前の無音時間 (秒)",
            "type": "number"
          },
          "speaker_uuid": {
            "title": "キャラクターの UUID",
            "type": "string"
          },
          "speedScale": {
            "description": "全体の話速を 0.5 ~ 2.0 の範囲で指定する (デフォルト: 1.0) 。\n2.0 で 2 倍速、0.5 で 0.5 倍速になる。",
            "title": "全体の話速",
            "type": "number"
          },
          "style_id": {
            "title": "スタイル ID",
            "type": "integer"
          },
          "tempoDynamicsScale": {
            "default": 1.0,
            "description": "話す速さの緩急の強弱を 0.0 ~ 2.0 の範囲で指定する (デフォルト: 1.0) 。\n値が大きいほどより早口で生っぽい抑揚がついた声になる。\nVOICEVOX ENGINE との互換性のため、未指定時はデフォルト値が適用される。",
            "title": "全体のテンポの緩急 (AivisSpeech Engine 固有のフィールド)",
            "type": "number"
          },
          "volumeScale": {
            "description": "全体の音量を 0.0 ~ 2.0 の範囲で指定する (デフォルト: 1.0) 。\n値が大きいほど大きな声になる。",
            "title": "全体の音量",
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "speaker_uuid",
          "style_id",
          "speedScale",
          "intonationScale",
          "pitchScale",
          "volumeScale",
          "prePhonemeLength",
          "postPhonemeLength"
        ],
        "title": "Preset",
        "type": "object"
      },
...
      "Speaker": {
        "description": "キャラクター情報",
        "properties": {
          "name": {
            "description": "名前",
            "title": "Name",
            "type": "string"
          },
          "speaker_uuid": {
            "description": "キャラクターの UUID",
            "title": "Speaker Uuid",
            "type": "string"
          },
          "styles": {
            "description": "スタイルの一覧",
            "items": {
              "$ref": "#/components/schemas/SpeakerStyle"
            },
            "title": "Styles",
            "type": "array"
          },
          "supported_features": {
            "$ref": "#/components/schemas/SpeakerSupportedFeatures",
            "description": "キャラクターの対応機能"
          },
          "version": {
            "description": "キャラクターのバージョン",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "name",
          "speaker_uuid",
          "styles",
          "version"
        ],
        "title": "Speaker",
        "type": "object"
      },
      "SpeakerInfo": {
        "description": "キャラクターの追加情報",
        "properties": {
          "policy": {
            "description": "policy.md",
            "title": "Policy",
            "type": "string"
          },
          "portrait": {
            "description": "アイコンを base64 エンコードしたもの、あるいは URL (「立ち絵画像を base64 エンコードしたもの」ではない点で VOICEVOX ENGINE と異なる)",
            "title": "Portrait",
            "type": "string"
          },
          "style_infos": {
            "description": "スタイルの追加情報",
            "items": {
              "$ref": "#/components/schemas/StyleInfo"
            },
            "title": "Style Infos",
            "type": "array"
          }
        },
        "required": [
          "policy",
          "portrait",
          "style_infos"
        ],
        "title": "SpeakerInfo",
        "type": "object"
      },
      "SpeakerStyle": {
        "description": "キャラクターのスタイル情報",
        "properties": {
          "id": {
            "description": "スタイル ID",
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "description": "スタイル名",
            "title": "Name",
            "type": "string"
          },
          "type": {
            "default": "talk",
            "description": "スタイルの種類。talk: 音声合成クエリの作成と音声合成が可能。singing_teacher: 歌唱音声合成用のクエリの作成が可能。frame_decode: 歌唱音声合成が可能。sing: 歌唱音声合成用のクエリの作成と歌唱音声合成が可能。",
            "enum": [
              "talk",
              "singing_teacher",
              "frame_decode",
              "sing"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "id"
        ],
        "title": "SpeakerStyle",
        "type": "object"
      },
      "SpeakerSupportedFeatures": {
        "description": "キャラクターの対応機能の情報",
        "properties": {
          "permitted_synthesis_morphing": {
            "default": "ALL",
            "description": "モーフィング機能への対応。'ALL' は「全て許可」、'SELF_ONLY' は「同じキャラクター内でのみ許可」、'NOTHING' は「全て禁止」",
            "enum": [
              "ALL",
              "SELF_ONLY",
              "NOTHING"
            ],
            "title": "Permitted Synthesis Morphing",
            "type": "string"
          }
        },
        "title": "SpeakerSupportedFeatures",
        "type": "object"
      },
      "StyleInfo": {
        "description": "スタイルの追加情報",
        "properties": {
          "icon": {
            "description": "このスタイルのアイコンを base64 エンコードしたもの、あるいは URL",
            "title": "Icon",
            "type": "string"
          },
          "id": {
            "description": "スタイル ID",
            "title": "Id",
            "type": "integer"
          },
          "portrait": {
            "description": "AivisSpeech Engine では常に None を返す (「このスタイルの立ち絵画像を base64 エンコードしたもの」ではない点で VOICEVOX ENGINE と異なる)",
            "title": "Portrait",
            "type": "string"
          },
          "voice_sample_transcripts": {
            "default": [],
            "description": "ボイスサンプルの書き起こしテキスト (voice_samples の配列インデックスと対応し、存在しない場合は空文字列)",
            "items": {
              "type": "string"
            },
            "title": "Voice Sample Transcripts",
            "type": "array"
          },
          "voice_samples": {
            "description": "ボイスサンプルの音声データを base64 エンコードしたもの、あるいは URL",
            "items": {
              "type": "string"
            },
            "title": "Voice Samples",
            "type": "array"
          }
        },
        "required": [
          "id",
          "icon",
          "voice_samples"
        ],
        "title": "StyleInfo",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AivisSpeech の音声合成エンジンです。",
    "title": "AivisSpeech Engine",
    "version": "latest"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/add_preset": {
      "post": {
        "description": "新しいプリセットを追加します。",
        "operationId": "add_preset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Preset",
                "description": "新しいプリセット。プリセットIDが既存のものと重複している場合は、新規のプリセットIDが採番されます。"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Add Preset Add Preset Post",
                  "type": "integer"
                }
              }
            },
            "description": "追加したプリセットのプリセットID"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "新しいプリセットを追加する",
        "tags": [
          "プリセット"
        ]
      }
    },
...
    "/aivm_models/install": {
      "post": {
        "description": "音声合成モデルをインストールします。\nファイルからインストールする場合は `file` を指定してください。\nURL からインストールする場合は `url` を指定してください。",
        "operationId": "install_model",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_install_model_aivm_models_install_post"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "音声合成モデルをインストールする",
        "tags": [
          "音声合成モデル管理"
        ]
      }
    },
    "/aivm_models/{aivm_uuid}": {
      "get": {
        "description": "指定された音声合成モデルの情報を取得します。",
        "operationId": "get_aivm_info",
        "parameters": [
          {
            "description": "音声合成モデルの UUID",
            "in": "path",
            "name": "aivm_uuid",
            "required": true,
            "schema": {
              "description": "音声合成モデルの UUID",
              "title": "Aivm Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AivmInfo"
                }
              }
            },
            "description": "指定された音声合成モデルの情報"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "指定された音声合成モデルの情報を取得する",
        "tags": [
          "音声合成モデル管理"
        ]
      }
    },
    "/aivm_models/{aivm_uuid}/load": {
      "post": {
        "description": "指定された音声合成モデルをロードします。すでにロード済みの場合は何も行われません。\n実行しなくても他の API は利用できますが、音声合成の初回実行時に時間がかかることがあります。",
        "operationId": "load_model",
        "parameters": [
          {
            "description": "音声合成モデルの UUID",
            "in": "path",
            "name": "aivm_uuid",
            "required": true,
            "schema": {
              "description": "音声合成モデルの UUID",
              "title": "Aivm Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "指定された音声合成モデルをロードする",
        "tags": [
          "音声合成モデル管理"
        ]
      }
    },
    "/aivm_models/{aivm_uuid}/uninstall": {
      "delete": {
        "description": "指定された音声合成モデルをアンインストールします。",
        "operationId": "uninstall_model",
        "parameters": [
          {
            "description": "音声合成モデルの UUID",
            "in": "path",
            "name": "aivm_uuid",
            "required": true,
            "schema": {
              "description": "音声合成モデルの UUID",
              "title": "Aivm Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "指定された音声合成モデルをアンインストールする",
        "tags": [
          "音声合成モデル管理"
        ]
      }
    },
    "/aivm_models/{aivm_uuid}/unload": {
      "post": {
        "description": "指定された音声合成モデルをアンロードします。",
        "operationId": "unload_model",
        "parameters": [
          {
            "description": "音声合成モデルの UUID",
            "in": "path",
            "name": "aivm_uuid",
            "required": true,
            "schema": {
              "description": "音声合成モデルの UUID",
              "title": "Aivm Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "指定された音声合成モデルをアンロードする",
        "tags": [
          "音声合成モデル管理"
        ]
      }
    },
    "/aivm_models/{aivm_uuid}/update": {
      "post": {
        "description": "AivisHub から指定された音声合成モデルの一番新しいバージョンをダウンロードし、\nインストール済みの音声合成モデルへ上書き更新します。",
        "operationId": "update_model",
        "parameters": [
          {
            "description": "音声合成モデルの UUID",
            "in": "path",
            "name": "aivm_uuid",
            "required": true,
            "schema": {
              "description": "音声合成モデルの UUID",
              "title": "Aivm Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "指定された音声合成モデルを更新する",
        "tags": [
          "音声合成モデル管理"
        ]
      }
    },
...
    "/delete_preset": {
      "post": {
        "description": "既存のプリセットを削除します。",
        "operationId": "delete_preset",
        "parameters": [
          {
            "description": "削除するプリセットのプリセットID",
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "description": "削除するプリセットのプリセットID",
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "既存のプリセットを削除する",
        "tags": [
          "プリセット"
        ]
      }
    },
...
    "/update_preset": {
      "post": {
        "description": "既存のプリセットを更新します。",
        "operationId": "update_preset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Preset",
                "description": "更新するプリセット。プリセットIDが更新対象と一致している必要があります。"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Update Preset Update Preset Post",
                  "type": "integer"
                }
              }
            },
            "description": "更新したプリセットのプリセットID"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "既存のプリセットを更新する",
        "tags": [
          "プリセット"
        ]
      }
    },
...
  }
}